Simplicity does not precede complexity, but follows it.

+++ backend/manage_db/delete_database.php
@@ -4,7 +4,15 @@ require_once('../common/connection.php');
 $db = new DbConnect($admin, $pass);
 
 $db_name = $_GET['db_name'];
-$sql = "DROP DATABASE $db_name";
+
+if ($db_name === in_array($restricted_names)){
+$restricted_names = [bill, bob, mysql, performance_schema, sys, information_schema];
+// send message to say delete stopped.
+} else {
+  $sql = "DROP DATABASE $db_name";
+//  go through command and delete
+}
+
 
 $result = $db->conn->query($sql);
